+++ pkg/webui/console/lib/feature-checks.js
@@ -109,6 +109,10 @@ export const mayEditApplicationDeviceKeys = {
   rightsSelector: selectApplicationRights,
   check: rights => rights.includes('RIGHT_APPLICATION_DEVICES_WRITE_KEYS'),
 }
+export const maySendUplink = {
+  rightsSelector: selectApplicationRights,
+  check: rights => rights.includes('RIGHT_APPLICATION_TRAFFIC_UP_WRITE') && asEnabled,
+}
 export const mayScheduleDownlinks = {
   rightsSelector: selectApplicationRights,
   check: rights => rights.includes('RIGHT_APPLICATION_TRAFFIC_DOWN_WRITE') && asEnabled,
@@ -220,4 +224,8 @@ export const mayViewGateways = {
   rightsSelector: state => [...selectUserRights(state), ...selectOrganizationRights(state)],
   check: rights =>
     mayViewApplicationsOfUser.check(rights) || mayViewApplicationsOfOrganization.check(rights),
+}
+export const mayWriteTraffic = {
+  rightsSelector: selectApplicationRights,
+  check: rights => mayScheduleDownlinks.check(rights) || maySendUplink.check(rights),
 }