+++ pkg/webui/console/views/device/index.js
@@ -30,7 +30,7 @@ import DeviceTitleSection from '@console/containers/device-title-section'
 
 import DeviceData from '@console/views/device-data'
 import DeviceGeneralSettings from '@console/views/device-general-settings'
-import DeviceMessages from '@console/views/device-messages'
+import DeviceMessaging from '@console/views/device-messaging'
 import DeviceLocation from '@console/views/device-location'
 import DevicePayloadFormatters from '@console/views/device-payload-formatters'
 import DeviceClaimAuthenticationCode from '@console/views/device-claim-authentication-code'
@@ -44,6 +44,7 @@ import sharedMessages from '@ttn-lw/lib/shared-messages'
 import {
   mayReadApplicationDeviceKeys,
   mayScheduleDownlinks,
+  maySendUplink,
   checkFromState,
 } from '@console/lib/feature-checks'
 
@@ -70,6 +71,7 @@ import style from './device.styl'
       device,
       mayReadKeys: checkFromState(mayReadApplicationDeviceKeys, state),
       mayScheduleDownlinks: checkFromState(mayScheduleDownlinks, state),
+      maySendUplink: checkFromState(maySendUplink, state),
       fetching: selectDeviceFetching(state),
       error: selectDeviceError(state),
     }
@@ -130,12 +132,13 @@ import style from './device.styl'
 @withEnv
 export default class Device extends React.Component {
   static propTypes = {
+    appId: PropTypes.string.isRequired,
     devId: PropTypes.string.isRequired,
     device: PropTypes.device.isRequired,
     env: PropTypes.env,
     location: PropTypes.location.isRequired,
-    match: PropTypes.match.isRequired,
     mayScheduleDownlinks: PropTypes.bool.isRequired,
+    maySendUplink: PropTypes.bool.isRequired,
     stopStream: PropTypes.func.isRequired,
   }
 
@@ -152,13 +155,12 @@ export default class Device extends React.Component {
   render() {
     const {
       location: { pathname },
-      match: {
-        params: { appId },
-      },
+      appId,
       devId,
       device,
       env: { siteName },
       mayScheduleDownlinks,
+      maySendUplink,
     } = this.props
     const {
       name,
@@ -178,6 +180,9 @@ export default class Device extends React.Component {
     const asConfig = selectAsConfig()
     const hasAs =
       asConfig.enabled && application_server_address === getHostnameFromUrl(asConfig.base_url)
+    const hideMessaging = !hasAs || !(mayScheduleDownlinks || maySendUplink)
+    const hidePayloadFormatters = !hasAs
+    const hideClaiming = !hasJs
 
     const basePath = `/applications/${appId}/devices/${devId}`
 
@@ -185,15 +190,19 @@ export default class Device extends React.Component {
     const payloadFormattersLink = pathname.startsWith(`${basePath}/payload-formatters`)
       ? pathname
       : `${basePath}/payload-formatters`
+    const messagingLink = pathname.startsWith(`${basePath}/messaging`)
+      ? pathname
+      : `${basePath}/messaging`
 
     const tabs = [
       { title: sharedMessages.overview, name: 'overview', link: basePath },
       { title: sharedMessages.liveData, name: 'data', link: `${basePath}/data` },
       {
-        title: sharedMessages.messages,
-        name: 'messages',
-        link: `${basePath}/messages`,
-        hidden: !mayScheduleDownlinks,
+        title: sharedMessages.messaging,
+        name: 'messaging',
+        exact: false,
+        link: messagingLink,
+        hidden: hideMessaging,
       },
       { title: sharedMessages.location, name: 'location', link: `${basePath}/location` },
       {
@@ -201,13 +210,13 @@ export default class Device extends React.Component {
         name: 'develop',
         link: payloadFormattersLink,
         exact: false,
-        hidden: !hasAs,
+        hidden: hidePayloadFormatters,
       },
       {
         title: sharedMessages.claiming,
         name: 'claim-auth-code',
         link: `${basePath}/claim-auth-code`,
-        hidden: !hasJs,
+        hidden: hideClaiming,
       },
       {
         title: sharedMessages.generalSettings,
@@ -233,13 +242,13 @@ export default class Device extends React.Component {
         <Switch>
           <Route exact path={basePath} component={DeviceOverview} />
           <Route exact path={`${basePath}/data`} component={DeviceData} />
-          <Route exact path={`${basePath}/messages`} component={DeviceMessages} />
+          {!hideMessaging && <Route path={`${basePath}/messaging`} component={DeviceMessaging} />}
           <Route exact path={`${basePath}/location`} component={DeviceLocation} />
           <Route exact path={`${basePath}/general-settings`} component={DeviceGeneralSettings} />
-          {hasAs && (
+          {!hidePayloadFormatters && (
             <Route path={`${basePath}/payload-formatters`} component={DevicePayloadFormatters} />
           )}
-          {hasJs && (
+          {!hideClaiming && (
             <Route path={`${basePath}/claim-auth-code`} component={DeviceClaimAuthenticationCode} />
           )}
           <NotFoundRoute />